Problem: Since we removed the previous migrate command we lost the capability of quickly prototyping applications assembling the code on our orbs.
Solution: Bring the functionality that calls assemble_schema back to a
command called assemble so the functionality is clearly tied to the
command name and we can still access it.
This is useful as a primitive for development even if it won't be commonly used once our
schema evolution command is in place.
